requirements.txt
pymupdf
ebooklib
beautifulsoup4
pytesseract
pillow
rich
pytest
readchar